Moved the classifier command into a subcommand of the train command.
